Honor "orphan" for `showLinePoints`
`"orphan"` is now an option for `showLinePoints` and will show points only for _orphan points_ of a constructed line, which is the points on a line that are isolated by there being undefined points immediately before and after. As such, "orphan" will only work in conjuction with `defined` having some setting that causes some parts of the line to be undefined.